+++ src/main/java/de/uib/configed/Configed.java
@@ -348,7 +348,8 @@ public final class Configed {
 	public static void main(CommandLine cmd) {
 		processArgs(cmd);
 
-		// Set timeout for login // seems that this timeout is not used if server is incorrect; setting timeout additionally in ServerFacade
+		// Set timeout for login
+		// seems that this timeout is not used if server is incorrect; setting timeout additionally in ServerFacade
 		System.setProperty("sun.net.client.defaultConnectTimeout", Globals.DEFAULT_TIMEOUT + "");
 		Logging.debug("configed: args recognized");