+++ src/main/java/de/uib/opsidatamodel/serverdata/OpsiServiceNOMPersistenceController.java
@@ -305,7 +305,6 @@ public class OpsiServiceNOMPersistenceController {
 		hostDataReloadHandler.setConfigDataService(configDataService);
 		hostDataReloadHandler.setGroupDataService(groupDataService);
 		hostDataReloadHandler.setHostInfoCollections(hostInfoCollections);
-		hostDataReloadHandler.setHealthDataService(healthDataService);
 		reloadDispatcher.registerHandler(ReloadEvent.HOST_DATA_RELOAD.toString(), hostDataReloadHandler);
 
 		ConfigOptionsDataReloadHandler configOptionsDataReloadHandler = new ConfigOptionsDataReloadHandler();